# backend/docs/swagger.yaml
definitions:
  entities.Lifestyle:
    enum:
    - not_active
    - active
    - sportive
    type: string
    x-enum-varnames:
    - NotActive
    - Active
    - Sportive
  entities.Want:
    enum:
    - lose_weight
    - build_muscle
    - stay_fit
    type: string
    x-enum-varnames:
    - LoseWeight
    - BuildMuscle
    - StayFit
  models.ErrorResponse:
    properties:
      '{route} error':
        example: '{method}: {crud}: error message'
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        example: Successfully {action}
        type: string
    type: object
...
  models.UserParamsCreateRequestModel:
    properties:
      height:
        maximum: 250
        minimum: 100
        type: integer
      lifestyle:
        enum:
        - not_active
        - active
        - sportive
        type: string
      photo:
        type: string
      targetCaloriesDaily:
        maximum: 10000
        minimum: 0
        type: integer
      targetWeight:
        maximum: 300
        minimum: 40
        type: number
      targetWorkoutsWeeks:
        maximum: 7
        minimum: 0
        type: integer
      wants:
        enum:
        - lose_weight
        - build_muscle
        - stay_fit
        type: string
    required:
    - height
    - lifestyle
    - targetCaloriesDaily
    - targetWeight
    - targetWorkoutsWeeks
    - wants
    type: object
  models.UserParamsResponseModel:
    properties:
      currentWeight:
        type: number
      height:
        type: integer
      lifestyle:
        $ref: '#/definitions/entities.Lifestyle'
      photo:
        type: string
      targetCaloriesDaily:
        type: integer
      targetWeight:
        type: number
      targetWorkoutsWeeks:
        type: integer
      wants:
        $ref: '#/definitions/entities.Want'
    type: object
  models.UserParamsUpdateRequestModel:
    properties:
      height:
        maximum: 250
        minimum: 100
        type: integer
      lifestyle:
        enum:
        - not_active
        - active
        - sportive
        type: string
      photo:
        type: string
      targetCaloriesDaily:
        maximum: 10000
        minimum: 0
        type: integer
      targetWeight:
        maximum: 300
        minimum: 40
        type: number
      targetWorkoutsWeeks:
        maximum: 7
        minimum: 0
        type: integer
      wants:
        enum:
        - lose_weight
        - build_muscle
        - stay_fit
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /user/params:
    delete:
      description: Удаляет дополнительные параметры пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление параметров пользователя
      tags:
      - User Params
    get:
      description: Получает дополнительные параметры пользователя (рост, возраст и
        т.д.)
      produces:
      - application/json
      responses:
        "200":
          description: Параметры пользователя
          schema:
            $ref: '#/definitions/models.UserParamsResponseModel'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение параметров пользователя
      tags:
      - User Params
    patch:
      consumes:
      - application/json
      description: Обновляет дополнительные параметры пользователя (рост, возраст
        и т.д.)
      parameters:
      - description: Данные параметров для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserParamsUpdateRequestModel'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Ошибка валидации или неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление параметров пользователя
      tags:
      - User Params
    post:
      consumes:
      - application/json
      description: Создает дополнительные параметры пользователя (рост, возраст и
        т.д.)
      parameters:
      - description: Данные параметров для создания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserParamsCreateRequestModel'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное создание
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Ошибка валидации или неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание параметров пользователя
      tags:
      - User Params
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...